+++ ui/packages/shared/profile/src/ProfileIcicleGraph/IcicleGraph/index.tsx
@@ -26,6 +26,8 @@ import useColoredGraph from './useColoredGraph';
 
 interface IcicleGraphProps {
   graph: Flamegraph;
+  total: number;
+  filtered: number;
   sampleUnit: string;
   width?: number;
   curPath: string[];
@@ -35,6 +37,8 @@ interface IcicleGraphProps {
 
 export const IcicleGraph = memo(function IcicleGraph({
   graph,
+  total,
+  filtered,
   width,
   setCurPath,
   curPath,
@@ -57,12 +61,13 @@ export const IcicleGraph = memo(function IcicleGraph({
     }
   }, [width, coloredGraph]);
 
-  const total = useMemo(() => parseFloat(coloredGraph.total), [coloredGraph.total]);
   const xScale = useMemo(() => {
     if (width === undefined) {
       return () => 0;
     }
-    return scaleLinear().domain([0, total]).range([0, width]);
+    return scaleLinear()
+      .domain([0, Number(total)])
+      .range([0, width]);
   }, [total, width]);
 
   if (coloredGraph.root === undefined || width === undefined) {
@@ -75,6 +80,7 @@ export const IcicleGraph = memo(function IcicleGraph({
       <GraphTooltip
         unit={sampleUnit}
         total={total}
+        totalUnfiltered={total + filtered}
         contextElement={svg.current}
         strings={coloredGraph.stringTable}
         mappings={coloredGraph.mapping}